+++ OUDS/Core/Themes/Orange/Tests/OrangeThemeTests.swift
@@ -52,6 +52,10 @@ struct OrangeThemeTests {
         #expect(theme.charts is OrangeThemeColorChartSemanticTokensProvider)
     }
 
+    @Test func defaultEffectemanticTokensProvider() throws {
+        #expect(theme.effects is OrangeThemeEffectSemanticTokensProvider)
+    }
+
     @Test func defaultElevationSemanticTokensProvider() throws {
         #expect(theme.elevations is OrangeThemeElevationSemanticTokensProvider)
     }