+++ OUDS/Core/Themes/OrangeBusinessTools/Tests/OrangeBusinessToolsThemeTests.swift
@@ -52,6 +52,10 @@ struct OrangeBusinessToolsThemeTests {
         #expect(theme.charts is OrangeBusinessToolsThemeColorChartSemanticTokensProvider)
     }
 
+    @Test func defaultEffectSemanticTokensProvider() throws {
+        #expect(theme.effects is OrangeBusinessToolsThemeEffectSemanticTokensProvider)
+    }
+
     @Test func defaultElevationSemanticTokensProvider() throws {
         #expect(theme.elevations is OrangeBusinessToolsThemeElevationSemanticTokensProvider)
     }