+++ OUDS/Core/ThemesContract/Sources/OUDSTheme/OUDSTheme.swift
@@ -58,6 +58,9 @@ open class OUDSTheme: @unchecked Sendable {
     public let charts: AllColorChartSemanticTokensProvider!
     // swiftlint:enable implicitly_unwrapped_optional
 
+    /// All effect semantic tokens exposed in one object
+    public let effects: AllEffectSemanticTokensProvider
+
     /// All elevation semantic tokens exposed in one object
     public let elevations: AllElevationSemanticTokensProvider
 
@@ -165,6 +168,7 @@ open class OUDSTheme: @unchecked Sendable {
     ///    - colors: All semantic tokens of colors
     ///    - colorModes: All semantic tokens of color modes
     ///    - charts: All semantic tokens of color charts if the theme have some, otherwise nil (by default is nil)
+    ///    - effects: All semantic tokens of effects
     ///    - elevations: All semantic tokens of elevations
     ///    - fonts: All semantic tokens of fonts
     ///    - grids: All semantic tokens of grids
@@ -199,6 +203,7 @@ open class OUDSTheme: @unchecked Sendable {
                 colors: AllColorSemanticTokensProvider,
                 colorModes: AllColorModeSemanticTokensProvider,
                 charts: AllColorChartSemanticTokensProvider? = nil,
+                effects: AllEffectSemanticTokensProvider,
                 elevations: AllElevationSemanticTokensProvider,
                 fonts: AllFontSemanticTokensProvider,
                 grids: AllGridSemanticTokensProvider,
@@ -237,6 +242,7 @@ open class OUDSTheme: @unchecked Sendable {
         self.colorModes = colorModes
         self.charts = charts
         self.elevations = elevations
+        self.effects = effects
         self.fonts = fonts
         self.grids = grids
         self.opacities = opacities