+++ OUDS/Core/Tokens/SemanticTokens/Sources/TypeAliases/EffectSemanticTokens+Aliases.swift
@@ -1,0 +1,17 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import OUDSTokensRaw
+
+/// Basically a semantic token for effect is a raw token for effect, to keep grammar clean and clear with design system grammar.
+public typealias EffectSemanticToken = EffectRawToken